+++ meteor/server/migration/upgrades/index.ts
@@ -1,0 +1,3 @@
+export * from './checkStatus'
+export * from './showStyleBase'
+export * from './studio'
